requirements.txt
Django==2.2.7
django-jalali==3.1.0
djangorestframework==3.10.3
jdatetime==3.6.2
pytz==2019.3
sqlparse==0.3.0
